+++ server/src/main/java/hupiat/intranet/server/assets/AssetRepository.java
@@ -2,6 +2,6 @@ package hupiat.intranet.server.assets;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AssetRepository extends JpaRepository<Asset, Short> {
+public interface AssetRepository extends JpaRepository<Asset, Long> {
 
 }
